+++ user/plugins/problems/CHANGELOG.md
@@ -1,3 +1,26 @@
+# v2.0.3
+## 05/09/2019
+
+1. [](#new)
+    * Code cleanup
+    * Pass `phpstan` tests
+    * Added `ru` and `uk` translations [#23](https://github.com/getgrav/grav-plugin-problems/pull/23)
+
+# v2.0.2
+## 12/16/2018
+
+1. [](#bugfix)
+    * Fixed an issue with checker not being initialized on Fatal Error
+
+# v2.0.1
+## 12/07/2018
+
+1. [](#new)
+    * Added support for admin reporting available in Grav 1.6
+1. [](#bugfix)
+    * Fixed issue with twig auto-escaping
+    * Fixed problems plugin potentially breaking CLI command if plugins get initialized
+
 # v2.0.0
 ## 09/30/2018